Attempting to use an enum for origins
This isn't called planets because it has the sun, and it isn't called CelestialObjects because that's too damn long. Also, Origins makes sense for the context in which they will be used: reference frame origins. However, I need to figure out how to update the values currently used when I'll be loading SPK files and interpolating their positions. Part of me thinks this will need to be in static, but I'm truly not sure that's a good design.